+++ beam-core/src/main/java/org/esa/beam/framework/datamodel/Product.java
@@ -1200,7 +1200,7 @@ public class Product extends ProductNode {
      * @return true, if so
      * @see #containsPixel(PixelPos)
      */
-    public boolean containsPixel(final float x, final float y) {
+    public boolean containsPixel(final double x, final double y) {
         return x >= 0.0f && x <= getSceneRasterWidth() &&
                 y >= 0.0f && y <= getSceneRasterHeight();
     }
@@ -1210,7 +1210,7 @@ public class Product extends ProductNode {
      *
      * @param pixelPos the pixel position, must not be null
      * @return true, if so
-     * @see #containsPixel(float, float)
+     * @see #containsPixel(double, double)
      */
     public boolean containsPixel(final PixelPos pixelPos) {
         return containsPixel(pixelPos.x, pixelPos.y);
