+++ beam-core/src/test/java/org/esa/beam/dataio/dimap/DimapProductWriterTest_WriteTiePointGrids.java
@@ -68,7 +68,7 @@ public class DimapProductWriterTest_WriteTiePointGrids extends TestCase {
     public void testWriteProductNodes_TiePointGrid() {
         Product product = new Product("name", BeamConstants.MERIS_FR_L1B_PRODUCT_TYPE_NAME,
                                       50, 25);
-        float[] expectedArray = getTiePointData(10, 5);
+        double[] expectedArray = getTiePointData(10, 5);
         TiePointGrid tiePointGrid = new TiePointGrid("name", 10, 5, 0, 0, 5, 5, expectedArray);
         product.addTiePointGrid(tiePointGrid);
 
@@ -91,8 +91,8 @@ public class DimapProductWriterTest_WriteTiePointGrids extends TestCase {
     ///   End Of Public
     ///////////////////////////////////////////////////////////////////////////
 
-    private float[] getTiePointData(int width, int height) {
-        float[] tiePoints = new float[width * height];
+    private double[] getTiePointData(int width, int height) {
+        double[] tiePoints = new double[width * height];
         for (int x = 0; x < tiePoints.length; x++) {
             tiePoints[x] = x * 1000;
         }
